requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
anaconda=custom=py37_0
ca-certificates=2018.03.07=0
certifi=2018.10.15=py37_0
freetype=2.9.1=h8a8886c_1
jpeg=9b=h024ee3a_2
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libpng=1.6.35=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.9=he85c1e1_2
ncurses=6.1=hf484d3e_0
olefile=0.46=py37_0
openssl=1.1.1=h7b6447c_0
pillow=5.3.0=py37h34e0f95_0
pip=18.1=py37_0
python=3.7.1=h0371630_3
readline=7.0=h7b6447c_5
setuptools=40.5.0=py37_0
sqlite=3.25.2=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.32.2=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=ha838bed_2
